requirements.txt
# --- Core API & backend ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

# --- Model frameworks ---
transformers==4.40.2
accelerate==0.30.1
sentence-transformers==2.7.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
timm==1.0.3
sentencepiece==0.2.0

# --- Retrieval & ranking ---
chromadb==0.4.24
faiss-cpu==1.7.4
rank-bm25==0.2.2

# --- Data & utils ---
pandas==2.2.2
pyarrow==16.1.0
numpy==1.26.4
tqdm==4.66.4
scikit-learn==1.4.2
requests==2.31.0

# --- Frontend / UI ---
streamlit==1.35.0
tabulate==0.9.0
# LangChain ecosystem
langchain==0.2.7
langchain-community==0.2.7
langchain-text-splitters==0.2.2
langsmith==0.1.84
